requirements.txt
flask==2.3.3
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pyyaml==6.0.1
docker==6.1.3